Package and test setup as well as problem #1
Set up the inital project structure including `problems` and `tests` packages. Also created the unit test class and solution class for Euler problem #1 - Multiples of 3 and 5. The object of the solution is to return the sum of all the natural numbers below N thatr are multiples of 3 or 5. The Project Euler problem required an answer for an input of 1000.